LayoutTests/http/tests/xmlhttprequest/redirections-and-user-headers-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8080/xmlhttprequest/resources/access-control-preflight-redirect.php?redirect=true&url=http://127.0.0.1:8000/xmlhttprequest/resources/access-control-preflight-redirect.php. Cross-origin redirection denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8080/xmlhttprequest/resources/access-control-preflight-redirect.php?redirect=true&url=http://localhost:8080/xmlhttprequest/resources/access-control-preflight-redirect.php. Cross-origin redirection denied by Cross-Origin Resource Sharing policy.

PASS Check headers after same-origin redirection to same-origin resource (simple request) 
PASS Check headers after same-origin redirection to same-origin resource (not simple request) 
PASS Check headers after same origin redirection to cross-origin resource (simple request) 
PASS Check headers after same origin redirection to cross-origin resource (not simple request) 
PASS Check headers after cross-origin redirection to same-origin resource (simple request) 
FAIL Check headers after cross-origin redirection to same-origin resource (not simple request) promise_test: Unhandled rejection with value: "Loading failure"
PASS Check headers after cross-origin redirection to cross-origin resource (simple request) 
FAIL Check headers after cross-origin redirection to cross-origin resource (not simple request) promise_test: Unhandled rejection with value: "Loading failure"
